// server/src/types/art.d.ts
export type Art = {
  objectID: number;
  isHighlight?: boolean;
  primaryImage?: string;
  additionalImages?: string[];
  department?: string;
  objectName?: string;
  culture?: string;
  period?: string;
  artistDisplayName?: string;
  artistAlphaSort?: string;
  artistDisplayBio?: string;
  artistNationality?: string;
  artistBeginDate?: string;
  artistEndDate?: string;
  objectDate?: string;
  dimensions?: string;
  city?: string;
  country?: string;
  repository?: string;
  note?: string;
};